# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SocialSphere API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete account
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/{userId}/follow:
    post:
      operationId: toggleFollow
      tags: [users]
      summary: Follow or unfollow user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResponse"
  /users/{userId}/followers:
    get:
      operationId: getFollowers
      tags: [users]
      summary: Get followers list
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/{userId}/following:
    get:
      operationId: getFollowing
      tags: [users]
      summary: Get following list
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Following
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/suggestions/list:
    get:
      operationId: getSuggestedUsers
      tags: [users]
      summary: Get suggested users to follow
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Suggested users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"

  # POSTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: string
        profilePicture:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        isFollowing:
          type: boolean
        followersCount:
          type: integer
      required: [id, username, displayName, isFollowing, followersCount]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        profilePicture:
          type: string
          nullable: true
        coverPhoto:
          type: string
          nullable: true
        followersCount:
          type: integer
        followingCount:
          type: integer
        postsCount:
          type: integer
        isFollowing:
          type: boolean
        isVerified:
          type: boolean
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
      required: [id, username, email, displayName, followersCount, followingCount, postsCount, isFollowing, isVerified, role, createdAt]
    UpdateProfileBody:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        profilePicture:
          type: string
        coverPhoto:
          type: string
        username:
          type: string
    FollowResponse:
      type: object
      properties:
        following:
          type: boolean
        followersCount:
          type: integer
      required: [following, followersCount]

    # POSTS
...